% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{check_db_creds}
\alias{check_db_creds}
\title{Check that the contents of the provided .yml credentials file contains valid
database credentials.}
\usage{
check_db_creds(db_creds, check_type = "warning")
}
\arguments{
\item{db_creds}{A list containing database credentials}

\item{check_type}{A character indicating whether errors found on checking should
induce a warning or a stop.}
}
\value{
A list containing the .yml file contents
}
\description{
Check that the contents of the provided .yml credentials file contains valid
database credentials.
}
