% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{create_connection}
\alias{create_connection}
\title{Create an Oracle database connection}
\usage{
create_connection(db_name, config_path = NULL)
}
\arguments{
\item{db_name}{A character string specifying the name of the database}

\item{config_path}{A character string specifying the file path for the .yml
file containing database credentials (optional)}
}
\value{
A ROracle dbconnect object.
}
\description{
Create an Oracle database connection
}
