% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_config_path}
\alias{get_config_path}
\title{Read config-path from system environment, and return a descriptive error if not found.}
\usage{
get_config_path()
}
\value{
A character string
}
\description{
Read config-path from system environment, and return a descriptive error if not found.
}
