% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{list_connections}
\alias{list_connections}
\title{List Oracle database connections available}
\usage{
list_connections(config_path = NULL)
}
\arguments{
\item{config_path}{A character string specifying the file path for the .yml
file containing database credentials (optional)}
}
\value{
A character vector of database names
}
\description{
List Oracle database connections available
}
