% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{read_db_creds}
\alias{read_db_creds}
\title{Read .yml configuration file}
\usage{
read_db_creds(db_name = NULL, config_path = NULL,
  check_type = "warning")
}
\arguments{
\item{config_path}{A character string specifying the file path for the .yml
file containing database credentials (optional)}
}
\value{
A list containing the .yml file contents
}
\description{
Read .yml configuration file
}
