% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{considerate_qry}
\alias{considerate_qry}
\title{Considerately run a SQL query, closing the database connection on completion.}
\usage{
considerate_qry(
  db_name,
  qry = NULL,
  qry_path = NULL,
  config_path = NULL,
  driver = NULL,
  ...
)
}
\arguments{
\item{db_name}{A list containing database credentials}

\item{qry}{A character string containing a SQL query}

\item{qry_path}{A character string containing a path to a .sql file}

\item{config_path}{Optional character string with the location to the config
file with database connection details}

\item{driver}{Optional character string specifying the database driver to use (currently
supports 'odbc' (default) or 'ROracle'.}

\item{...}{Optional arguments to be passed to write.csv}
}
\value{
A data frame containing records fetched by a SQL query
}
\description{
Considerately run a SQL query, closing the database connection on completion.
}
