% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{create_connection}
\alias{create_connection}
\title{Create an Oracle database connection}
\usage{
create_connection(db_name, config_path = NULL, driver = NULL)
}
\arguments{
\item{db_name}{A character string specifying the name of the database}

\item{config_path}{A character string specifying the file path for the .yml
file containing database credentials (optional)}

\item{driver}{Optional character string specifying the database driver to use (currently
supports 'odbc' (default) or 'ROracle'.}
}
\value{
A dbconnect object.
}
\description{
Create an Oracle database connection
}
